Move HandlePositionProvider to common
