+++ compose/foundation/foundation/src/commonMain/kotlin/androidx/compose/foundation/text/selection/SelectionHandles.kt
@@ -18,11 +18,19 @@ package androidx.compose.foundation.text.selection
 
 import androidx.compose.foundation.text.Handle
 import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.takeOrElse
 import androidx.compose.ui.semantics.SemanticsPropertyKey
 import androidx.compose.ui.text.style.ResolvedTextDirection
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.IntRect
+import androidx.compose.ui.unit.IntSize
+import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.round
+import androidx.compose.ui.window.PopupPositionProvider
 
 internal val HandleWidth = 25.dp
 internal val HandleHeight = 25.dp
@@ -104,4 +112,71 @@ internal fun interface OffsetProvider {
  */
 internal fun getAdjustedCoordinates(position: Offset): Offset {
     return Offset(position.x, position.y - 1f)
+}
+
+/**
+ * This [PopupPositionProvider] for a selection handle. It will position the selection handle
+ * to the result of [positionProvider] in its anchor layout.
+ */
+internal class HandlePositionProvider(
+    private val alignment: Alignment,
+    private val positionProvider: OffsetProvider,
+    private val offset: Offset = Offset.Zero,
+) : PopupPositionProvider {
+
+    /**
+     * When Handle disappears, it starts reporting its position as [Offset.Unspecified]. Normally,
+     * Popup is dismissed immediately when its position becomes unspecified, but for one frame a
+     * position update might be requested by soon-to-be-destroyed Popup. In this case, report the
+     * last known position as there are no more updates. If the first ever position is provided as
+     * unspecified, start with [Offset.Zero] default.
+     */
+    private var prevPosition: Offset = Offset.Zero
+
+    override fun calculatePosition(
+        anchorBounds: IntRect,
+        windowSize: IntSize,
+        layoutDirection: LayoutDirection,
+        popupContentSize: IntSize
+    ): IntOffset {
+        val position = positionProvider.provide().takeOrElse { prevPosition } + offset
+        prevPosition = position
+
+        val adjustment = alignment.align(popupContentSize, IntSize.Zero, layoutDirection)
+        return anchorBounds.topLeft + position.round() + adjustment
+    }
+}
+
+/**
+ * Computes whether the handle's appearance should be left-pointing or right-pointing.
+ */
+internal fun isLeftSelectionHandle(
+    isStartHandle: Boolean,
+    direction: ResolvedTextDirection,
+    handlesCrossed: Boolean
+): Boolean {
+    return if (isStartHandle) {
+        isHandleLtrDirection(direction, handlesCrossed)
+    } else {
+        !isHandleLtrDirection(direction, handlesCrossed)
+    }
+}
+
+/**
+ * This method is to check if the selection handles should use the natural Ltr pointing
+ * direction.
+ * If the context is Ltr and the handles are not crossed, or if the context is Rtl and the handles
+ * are crossed, return true.
+ *
+ * In Ltr context, the start handle should point to the left, and the end handle should point to
+ * the right. However, in Rtl context or when handles are crossed, the start handle should point to
+ * the right, and the end handle should point to left.
+ */
+/*@VisibleForTesting*/
+private fun isHandleLtrDirection(
+    direction: ResolvedTextDirection,
+    areHandlesCrossed: Boolean
+): Boolean {
+    return direction == ResolvedTextDirection.Ltr && !areHandlesCrossed ||
+        direction == ResolvedTextDirection.Rtl && areHandlesCrossed
 }
